+++ src/entities/ImgBuffer_t.cpp
@@ -13,7 +13,7 @@
 using APTracer::Entities::ImgBuffer_t;
 using APTracer::Entities::Vec3f;
 
-ImgBuffer_t::ImgBuffer_t(unsigned int size_x, unsigned int size_y): size_x_(size_x), size_y_(size_y), updates_(0) {
+ImgBuffer_t::ImgBuffer_t(size_t size_x, size_t size_y): size_x_(size_x), size_y_(size_y), updates_(0) {
     img_ = new Vec3f[size_y_*size_x_];
 }
 
@@ -23,15 +23,15 @@ ImgBuffer_t::~ImgBuffer_t() {
 
 void ImgBuffer_t::reset() {
     updates_ = 0;
-    for (unsigned int j = 0; j < size_y_*size_x_; ++j) {
+    for (size_t j = 0; j < size_y_*size_x_; ++j) {
         img_[j] = Vec3f();
     }
 }
 
-void ImgBuffer_t::update(const Vec3f* img, unsigned int size_x, unsigned int size_y) {
+void ImgBuffer_t::update(const Vec3f* img, size_t size_x, size_t size_y) {
     updates_++;
 
-    for (unsigned int j = 0; j < size_y*size_x; ++j) {
+    for (size_t j = 0; j < size_y*size_x; ++j) {
         //img_[j][i] = img_[j][i] * (1.0 - 1.0/static_cast<double>(updates_)) + img[j][i]/static_cast<double>(updates_);
         img_[j] += img[j];
     }
@@ -41,19 +41,19 @@ void ImgBuffer_t::update() {
     ++updates_;
 }
 
-void ImgBuffer_t::update(const Vec3f &colour, unsigned int pos_x, unsigned int pos_y) {
+void ImgBuffer_t::update(const Vec3f &colour, size_t pos_x, size_t pos_y) {
     //img_[pos_y][pos_x] = img_[pos_y][pos_x] * (1.0 - 1.0/static_cast<double>(updates_)) + colour/static_cast<double>(updates_);
     img_[pos_y*size_x_ + pos_x] += colour;
 }
 
-void ImgBuffer_t::set(const Vec3f* img, unsigned int size_x, unsigned int size_y) {
+void ImgBuffer_t::set(const Vec3f* img, size_t size_x, size_t size_y) {
     updates_ = 1;
-    for (unsigned int j = 0; j < size_y*size_x; ++j) {
+    for (size_t j = 0; j < size_y*size_x; ++j) {
         img_[j] = img[j];
     }
 }
 
-void ImgBuffer_t::set(const Vec3f &colour, unsigned int pos_x, unsigned int pos_y) {
+void ImgBuffer_t::set(const Vec3f &colour, size_t pos_x, size_t pos_y) {
     img_[pos_y*size_x_ + pos_x] = colour;
 }
 
@@ -66,8 +66,8 @@ void ImgBuffer_t::write(std::string &filename) const {
     //constexpr double bit_multiplier = std::pow(2.0, bit_depth) - 1.0;
     constexpr double bit_multiplier = 65535.0; // 16 bits, with msvc std::pow is not constexpr :(
 
-    for (unsigned int j = 0; j < size_y_; ++j) {
-        for (unsigned int i = 0; i < size_x_; ++i) {
+    for (size_t j = 0; j < size_y_; ++j) {
+        for (size_t i = 0; i < size_x_; ++i) {
             Vec3f colour = img_[j*size_x_ + i]*update_mult;
             colour.clamp(0.0, 1.0);
             colour *= bit_multiplier;
@@ -89,8 +89,8 @@ void ImgBuffer_t::write(std::string &filename, double gammaind) const {
     //constexpr double bit_multiplier = std::pow(2.0, bit_depth) - 1.0;
     constexpr double bit_multiplier = 65535.0; // 16 bits, with msvc std::pow is not constexpr :(
 
-    for (unsigned int j = 0; j < size_y_; ++j) {
-        for (unsigned int i = 0; i < size_x_; ++i) {
+    for (size_t j = 0; j < size_y_; ++j) {
+        for (size_t i = 0; i < size_x_; ++i) {
             Vec3f colour = img_[j*size_x_ + i]*update_mult;
             colour.clamp(0.0, 1.0).pow_inplace(gammaind);
             colour *= bit_multiplier;
